requirements.txt
numpy==1.26.4
pandas==2.2.2
streamlit
plotly
matplotlib
pyspark
delta-spark